servers/replica/requirements.txt
arrow==1.1.1
astroid==2.8.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
flask==2.0.1
html5validator==0.4.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
grpcio==1.54.0
